+++ RockPaperScissors/RockPaperScissors/Controllers/main.swift
@@ -1,0 +1,9 @@
+//
+//  RockPaperScissors - main.swift
+//  Created by tacocat.
+//  Copyright © tastycode. All rights reserved.
+// 
+
+import Foundation
+
+RpsGame().play()
