+++ RockPaperScissors/RockPaperScissors/Views/Message.swift
@@ -1,5 +1,5 @@
 //
-//  Game.swift
+//  Message.swift
 //  RockPaperScissors
 //
 //  Created by 장우석 on 11/21/23.
@@ -7,13 +7,13 @@
 
 import Foundation
 
-enum RockPaperScissors {
+enum Message {
     case win
     case draw
     case lose
-    case menuText
-    case wrongText
-    case endText
+    case menu
+    case wrong
+    case end
     
     var result: String {
         switch self {
@@ -23,13 +23,12 @@ enum RockPaperScissors {
             return "졌습니다!"
         case .draw:
             return "비겼습니다!"
-        case .menuText:
+        case .menu:
             return "가위(1), 바위(2), 보(3)! <종료 : 0> :"
-        case .wrongText:
+        case .wrong:
             return "잘못된 입력입니다. 다시 시도해주세요."
-        case .endText:
+        case .end:
             return "게임종료"
         }
     }
 }
-
